+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added creation of sub model object starting from the full model and a sub traversal (https://github.com/robotology/idyntree/pull/1011).
+
 ### Fixed
 - Fix bug of `InverseKinematics::updateRotationTarget`  that did not actually changed the RotationWeight, but changed the PositionWeight instead (https://github.com/robotology/idyntree/pull/1012).
+
 
 ## [6.0.0] - 2022-07-19